Restructure file hierarchy
Rename Raindrops.csproj to RaindropsLibrary.csproj.
Rename RaindropFactors.cs to Raindrops.cs.
