+++ Raindrops/Raindrops.sln
@@ -3,9 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30309.148
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Raindrops", "Raindrops\Raindrops.csproj", "{0CA462BD-E511-46A1-BC31-B53EF0483C93}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RaindropsTests", "RaindropsTests\RaindropsTests.csproj", "{C07A0696-8425-4311-AAB3-69CB1CD98758}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RaindropsTests", "RaindropsTests\RaindropsTests.csproj", "{C07A0696-8425-4311-AAB3-69CB1CD98758}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RaindropsLibrary", "RaindropsLibrary\RaindropsLibrary.csproj", "{DDFD1B7D-C654-433D-BA8B-C8663BFB7EA0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,14 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0CA462BD-E511-46A1-BC31-B53EF0483C93}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0CA462BD-E511-46A1-BC31-B53EF0483C93}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0CA462BD-E511-46A1-BC31-B53EF0483C93}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0CA462BD-E511-46A1-BC31-B53EF0483C93}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C07A0696-8425-4311-AAB3-69CB1CD98758}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C07A0696-8425-4311-AAB3-69CB1CD98758}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C07A0696-8425-4311-AAB3-69CB1CD98758}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C07A0696-8425-4311-AAB3-69CB1CD98758}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DDFD1B7D-C654-433D-BA8B-C8663BFB7EA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DDFD1B7D-C654-433D-BA8B-C8663BFB7EA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DDFD1B7D-C654-433D-BA8B-C8663BFB7EA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DDFD1B7D-C654-433D-BA8B-C8663BFB7EA0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
